Add test data generation and user management utilities
